ref-cliend/salesforce_client/requirements.txt
grpcio
grpcio-tools
protobuf
google-auth
python-dotenv
requests
PyJWT
cryptography
fastavro==1.12.0
